Fix handling of skip_ansible_lint
This change modify task normalization to handle "tags" clause so that tags
are always stored as a list. This is to fix a problem about the "skip_ansible_lint"
tag. Below is the description of the problem:

Ansible-playbook allows "tags" clause to be either a scalar or a list.
That is, it accepts both of below:

  tags: skip_ansible_lint

  tags:
  - skip_ansible_lint

However, ansible-lint assumes only the list form and uses "in" operator
to see if a task has "skip_ansible_lint" or not. So if a task has a tag
"do_not_skip_ansible_lint" in the scalar form, what is performed is
"skip_ansible_lint" in "do_not_skip_ansible_lint" and it results in
True (lint is skipped). This is a different behavior from that of
"do_not_skip_ansible_lint" in the scalar form.